1. 每次写入SN都保存，不再判定是否为新SN 2. 保存SN时增加重复校验，去除重复的数据

+++ Tool/src/com/petkit/matetool/ui/D4/utils/D4Utils.java
@@ -250,6 +250,10 @@ public class D4Utils {
         if(device == null || !device.checkValid()) {
             throw  new RuntimeException("store D4 failed, " + (device == null ? "D4 is null !" : device.toString()));
         }
+        //去重
+        if (FileUtils.isFileContainsString(getStoreDeviceInfoFilePath(), device.getSn())) {
+            return;
+        }
 
         PetkitLog.d("store D4 info: " + device.generateMainJson(ageingResult));
         FileUtils.writeStringToFile(getStoreDeviceInfoFilePath(), device.generateMainJson(ageingResult) + ",", true);
